# platform-api/plugins/eventgateway/openapi.yaml
openapi: "3.0.3"
info:
  title: Event Gateway Plugin API
  version: "0.9"
  description: >
    WebSub and WebBroker API management endpoints provided by the
    eventgateway compile-time plugin.
paths:
...
  /api/v1/websub-apis/{apiId}/secrets:
    parameters:
      - name: apiId
        in: path
        required: true
        schema:
          type: string
    post:
      security:
        - OAuth2SecurityEventGateway:
            - ap:websub_api:hmac_secret:create
            - ap:websub_api:manage
      responses:
        '201':
          description: HMAC secret created successfully
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '409':
          $ref: '#/components/responses/Conflict'
        '500':
          $ref: '#/components/responses/InternalServerError'
        '503':
          $ref: '#/components/responses/ServiceUnavailable'
    get:
      security:
        - OAuth2SecurityEventGateway:
            - ap:websub_api:hmac_secret:read
            - ap:websub_api:manage
      responses:
        '200':
          description: HMAC secrets retrieved successfully
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /api/v1/websub-apis/{apiId}/secrets/{secretName}:
    parameters:
      - name: apiId
        in: path
        required: true
        schema:
          type: string
      - name: secretName
        in: path
        required: true
        schema:
          type: string
    delete:
      security:
        - OAuth2SecurityEventGateway:
            - ap:websub_api:hmac_secret:delete
            - ap:websub_api:manage
      responses:
        '204':
          description: HMAC secret deleted successfully
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /api/v1/websub-apis/{apiId}/secrets/{secretName}/regenerate:
    parameters:
      - name: apiId
        in: path
        required: true
        schema:
          type: string
      - name: secretName
        in: path
        required: true
        schema:
          type: string
    post:
      security:
        - OAuth2SecurityEventGateway:
            - ap:websub_api:hmac_secret:regenerate
            - ap:websub_api:manage
      responses:
        '200':
          description: HMAC secret regenerated successfully
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          $ref: '#/components/responses/Conflict'
        '500':
          $ref: '#/components/responses/InternalServerError'
...
components:
...
  responses:
    BadRequest:
      description: Bad request
    Unauthorized:
      description: Unauthorized
    NotFound:
      description: Not found
    Conflict:
      description: Conflict
...
    InternalServerError:
      description: Internal server error
    ServiceUnavailable:
      description: Service unavailable
  securitySchemes:
    OAuth2SecurityEventGateway:
      type: oauth2
      flows:
        clientCredentials:
          tokenUrl: /oauth2/token
          scopes: {}